tests/wpt/tests/content-security-policy/img-src/css-filter-blocked.tentative.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="img-src 'none'">
    <title>SVG in CSS filter: blocked as img</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
      async_test(t => {
        window.addEventListener('securitypolicyviolation',
                                t.step_func_done(e => {
          assert_equals(e.effectiveDirective, "img-src");
          assert_true(e.blockedURI.endsWith("filter.svg"));
        }));
      }, "Blocked CSS filter: SVG");
    </script>
    <link rel="stylesheet" href="resources/blue.css">
</head>
<body>
  <style>
    body {
      filter: url("filter.svg#bar");
    }
  </style>
</body>
</html>
